SDK-1755:Add abstract Yoti Exception
