+++ src/Exception/AmlException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yoti\Exception;
 
-class AmlException extends \Exception
+use Yoti\Exception\base\YotiException;
+
+class AmlException extends YotiException
 {
 }
