+++ src/Exception/ShareUrlException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yoti\Exception;
 
-class ShareUrlException extends \Exception
+use Yoti\Exception\base\YotiException;
+
+class ShareUrlException extends YotiException
 {
 }
